Fix CC violations
