+++ hailstorm-gem/lib/hailstorm/model/cluster.rb
@@ -217,9 +217,9 @@ class Hailstorm::Model::Cluster < ActiveRecord::Base
     # check if load generation is not stopped on any load agent and raise
     # exception accordingly
     agents_running = Hailstorm::Model::LoadAgent
-                       .joins(jmeter_plan: :project)
-                       .where(projects: { id: project.id }, load_agents: { active: true })
-                       .where('load_agents.jmeter_pid IS NOT NULL')
+                     .joins(jmeter_plan: :project)
+                     .where(projects: { id: project.id }, load_agents: { active: true })
+                     .where('load_agents.jmeter_pid IS NOT NULL')
 
     return cluster_instances if agents_running.empty?
