+++ hailstorm-web-client/e2e/features/support/pages/WizardBase.ts
@@ -1,0 +1,4 @@
+export class WizardBase {
+  get nextLink() { return $('=Next'); }
+  get nextButton() { return $('button*=Next'); }
+}
