+++ hailstorm-web-client/e2e/features/support/pages/WizardReview.ts
@@ -1,0 +1,11 @@
+import { WizardBase } from './WizardBase';
+
+export class WizardReview extends WizardBase {
+
+  get doneButton() { return $("button*=Done"); }
+
+  finalize() {
+    browser.waitUntil(() => this.doneButton.isExisting(), 20000);
+    this.doneButton.click();
+  }
+}
